+++ src/RecoveryServices/RecoveryServices.Backup.Providers/Providers/MabPsBackupProvider.cs
@@ -46,12 +46,17 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
             ServiceClientAdapter = serviceClientAdapter;
         }
 
-        public RestAzureNS.AzureOperationResponse EnableProtection()
+        public RestAzureNS.AzureOperationResponse<ProtectedItemResource> EnableProtection()
         {
             throw new NotImplementedException();
         }
 
-        public RestAzureNS.AzureOperationResponse DisableProtection()
+        public RestAzureNS.AzureOperationResponse<ProtectedItemResource> DisableProtection()
+        {
+            throw new NotImplementedException();
+        }
+
+        public RestAzureNS.AzureOperationResponse DisableProtectionWithDeleteData()
         {
             throw new NotImplementedException();
         }