+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added Sql server in Azure VM support
 
 ## Version 1.0.0
 * General availability of `Az.RecoveryServices` module